move single_batch_overlap.py, two_batch_overlap.py, operations.py, operations_strategy.py under a new folder srt/batch_overlap
