+++ python/sglang/srt/layers/moe/token_dispatcher/mooncake.py
@@ -19,7 +19,7 @@ from sglang.srt.layers.moe.utils import DeepEPMode
 from sglang.srt.utils import get_int_env_var
 
 if TYPE_CHECKING:
-    from sglang.srt.single_batch_overlap import CombineOverlapArgs
+    from sglang.srt.batch_overlap.single_batch_overlap import CombineOverlapArgs
 
 from enum import Enum, auto
